+++ jjapcloudver2/src/main/java/com/iucyh/jjapcloudimprove/dto/music/query/MusicSimpleDto.java
@@ -10,13 +10,13 @@ public class MusicSimpleDto {
 
     private final String publicId;
     private final String title;
-    private final Integer playTime;
+    private final Long playTime;
     private final Integer viewCount;
     private final LocalDateTime createdAt;
     private final LocalDateTime updatedAt;
 
     @QueryProjection
-    public MusicSimpleDto(String publicId, String title, Integer playTime, Integer viewCount, LocalDateTime createdAt, LocalDateTime updatedAt) {
+    public MusicSimpleDto(String publicId, String title, Long playTime, Integer viewCount, LocalDateTime createdAt, LocalDateTime updatedAt) {
         this.publicId = publicId;
         this.title = title;
         this.playTime = playTime;
